+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import {
   Switch, Route, Redirect,
 } from 'react-router-dom';
-import Forms from './components/forms';
+import SimpleCalc from './components/SimpleCalc';
 import DistanceHome from './containers/DistanceHome';
 import Nav from './components/Nav/Nav';
 import Footer from './components/Footer/Footer';
@@ -13,7 +13,7 @@ function App() {
       <Nav />
 
       <Switch>
-        <Route path="/manual" exact component={Forms} />
+        <Route path="/manual" exact component={SimpleCalc} />
         <Route path="/" exact component={DistanceHome} />
         <Redirect to="/" />
       </Switch>
