+++ src/components/SimpleCalc.js
